pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "github-mcp-server"
version = "1.1.1"
description = "A comprehensive Model Context Protocol server for GitHub integration"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "MCP Labs", email = "licensing@mcplabs.co.uk" }]
license = { text = "AGPL-3.0-or-later" }
keywords = ["mcp", "model-context-protocol", "github", "api", "assistant", "claude"]
dependencies = [
  "mcp>=1.0.0",
  "httpx>=0.25.0",
  "pydantic>=2.0.0",
  "PyJWT>=2.8.0",
]

[project.urls]
"Bug Reports" = "https://github.com/crypto-ninja/github-mcp-server/issues"
"Source" = "https://github.com/crypto-ninja/github-mcp-server"
"Documentation" = "https://github.com/crypto-ninja/github-mcp-server/blob/main/README.md"
"Website" = "https://mcplabs.co.uk"
"Licensing" = "https://github.com/crypto-ninja/github-mcp-server/blob/main/LICENSING.md"

[project.scripts]
github-mcp = "github_mcp:main"


